add application.proprietes

+++ fut-sexta/src/main/java/com/fut_sexta/fut_sexta/DTO/ArtilheiroDTO.java
@@ -1,4 +1,4 @@
 package com.fut_sexta.fut_sexta.DTO;
 
-public record ArtilheiroDTO(String name, int goals) {
+public record ArtilheiroDTO(String name, Long goals) {
 }
